feat: updated store along with functions
