+++ apps/@sparrow-desktop/src/pages/workspace-page/sub-pages/GraphqlExplorerPage/GraphqlExplorerPage.ViewModel.ts
@@ -3,7 +3,8 @@ import {
   DecodeGraphql,
   ReduceAuthHeader,
 } from "@sparrow/workspaces/features/graphql-explorer/utils";
-import { createDeepCopy, moveNavigation, startLoading, stopLoading } from "@sparrow/common/utils";
+import { createDeepCopy, moveNavigation } from "@sparrow/common/utils";
+import { startLoading,stopLoading } from "../../../../../../../packages/@sparrow-common/src/store";
 import { CompareArray, Debounce } from "@sparrow/common/utils";
 
 // ---- DB
