+++ apps/@sparrow-desktop/src/pages/workspace-page/sub-pages/RestExplorer/RestExplorerPage.ViewModel.ts
@@ -6,7 +6,8 @@ import {
   ReduceAuthHeader,
   ReduceAuthParameter,
 } from "@sparrow/workspaces/features/rest-explorer/utils";
-import { createDeepCopy, moveNavigation, startLoading, stopLoading } from "@sparrow/common/utils";
+import { createDeepCopy, moveNavigation, } from "@sparrow/common/utils";
+import { startLoading,stopLoading } from "../../../../../../../packages/@sparrow-common/src/store";
 import {
   CompareArray,
   Debounce,
