+++ apps/@sparrow-desktop/src/pages/workspace-page/sub-pages/WebSocketExplorerPage/WebSocketExplorerPage.ViewModel.ts
@@ -4,7 +4,10 @@ import {
   ReduceQueryParams,
   DecodeWebsocket,
 } from "@sparrow/workspaces/features/socket-explorer/utils";
-import { createDeepCopy, moveNavigation, startLoading, stopLoading } from "@sparrow/common/utils";
+import { createDeepCopy, moveNavigation,
+  
+} from "@sparrow/common/utils";
+import { startLoading,stopLoading } from "../../../../../../../packages/@sparrow-common/src/store";
 import {
   CompareArray,
   Debounce,
