+++ apps/@sparrow-web/src/pages/workspace-page/sub-pages/RestExplorer/RestExplorerPage.ViewModel.ts
@@ -6,12 +6,11 @@ import {
   ReduceAuthHeader,
   ReduceAuthParameter,
 } from "@sparrow/workspaces/features/rest-explorer/utils";
+import { createDeepCopy, moveNavigation } from "@sparrow/common/utils";
 import {
-  createDeepCopy,
-  moveNavigation,
   startLoading,
   stopLoading,
-} from "@sparrow/common/utils";
+} from "../../../../../../../packages/@sparrow-common/src/store";
 import {
   CompareArray,
   Debounce,
