+++ apps/@sparrow-web/src/pages/workspace-page/sub-pages/SocketIoExplorerPage/SocketIoExplorerPage.ts
@@ -4,12 +4,11 @@ import {
   ReduceQueryParams,
   DecodeSocketio,
 } from "@sparrow/workspaces/features/socketio-explorer/utils";
+import { createDeepCopy, moveNavigation } from "@sparrow/common/utils";
 import {
-  createDeepCopy,
-  moveNavigation,
   startLoading,
   stopLoading,
-} from "@sparrow/common/utils";
+} from "../../../../../../../packages/@sparrow-common/src/store";
 import { CompareArray, Debounce } from "@sparrow/common/utils";
 
 // ---- DB
