+++ apps/@sparrow-web/src/pages/workspace-page/sub-pages/WebSocketExplorerPage/WebSocketExplorerPage.ViewModel.ts
@@ -4,12 +4,11 @@ import {
   ReduceQueryParams,
   DecodeWebsocket,
 } from "@sparrow/workspaces/features/socket-explorer/utils";
+import { createDeepCopy, moveNavigation } from "@sparrow/common/utils";
 import {
-  createDeepCopy,
-  moveNavigation,
   startLoading,
   stopLoading,
-} from "@sparrow/common/utils";
+} from "../../../../../../../packages/@sparrow-common/src/store";
 import {
   CompareArray,
   Debounce,
