--- packages/@sparrow-common/src/stores/index.ts
@@ -1,1 +1,0 @@
-export { loadingState } from "./loading";
