--- packages/@sparrow-common/src/stores/loading.ts
@@ -1,4 +1,0 @@
-import { writable } from "svelte/store";
-
-// Create the writable store with an initial empty object
-export const loadingState = writable<Map<string | number, boolean>>(new Map());
