+++ packages/@sparrow-common/src/utils/index.ts
@@ -19,7 +19,6 @@ export { handleDownloadResponse } from "./handle-downlod-rest";
 export { Sleep } from "./sleep";
 export { Base64Converter } from "./base64Converter";
 export { InitSavedRequestTab } from "./init-saved-request";
-export * from "./save-loading";
 export * from "./auth.helper";
 export * from "./common.helper";
 export * from "./conversion.helper";
