--- packages/@sparrow-common/src/utils/save-loading.ts
@@ -1,16 +1,0 @@
-import { loadingState } from "../stores";
-export const startLoading = (id: string) => {
-  loadingState.update((currentMap) => {
-    const updatedMap = new Map(currentMap);
-    updatedMap.set(id, true);
-    return updatedMap;
-  });
-};
-
-export const stopLoading = (id: string) => {
-  loadingState.update((currentMap) => {
-    const updatedMap = new Map(currentMap);
-    updatedMap.set(id, false);
-    return updatedMap;
-  });
-};
